+++ server/src/modules/jobs/jobs.ts
@@ -1,51 +1,56 @@
-import { IJob } from "shared/models/job.model";
+import { exec as nodeexec } from "node:child_process";
+import util from "node:util";
 
-import logger from "@/common/logger";
-import { sleep } from "@/common/utils/asyncUtils";
-import { getDbCollection } from "@/common/utils/mongodbUtils";
+import { IJob } from "shared/models/job.model";
+const exec = util.promisify(nodeexec);
+
 import {
   create as createMigration,
   status as statusMigration,
   up as upMigration,
 } from "@/modules/jobs/migrations/migrations";
 
+import logger from "../../common/logger";
 import { handleDocumentFileContent } from "../actions/documents.actions";
-import { createJob } from "../actions/job.actions";
 import { processMailingList } from "../actions/mailingLists.actions";
 import { createUser } from "../actions/users.actions";
+import { cronsInit, cronsScheduler } from "./crons_actions";
 import { recreateIndexes } from "./db/recreateIndexes";
 import { validateModels } from "./db/schemaValidation";
-import { executeJob } from "./executeJob";
+import { addJob, executeJob } from "./jobs_actions";
 import { clear } from "./seed/clear";
 import { seed } from "./seed/seed";
 
-export async function addJob(
-  {
-    name,
-    payload = {},
-    scheduled_at = new Date(),
-    sync = false,
-  }: Pick<IJob, "name"> &
-    Partial<Pick<IJob, "payload" | "scheduled_at" | "sync">>,
-  options: { runningLogs: boolean } = {
-    runningLogs: true,
-  }
-): Promise<number> {
-  const job = await createJob({
-    name,
-    payload,
-    scheduled_at,
-    sync,
-  });
-
-  if (sync && job) {
-    return runJob(job, options);
-  }
-
-  return 0;
+interface CronDef {
+  name: string;
+  cron_string: string;
+  handler: () => Promise<number>;
 }
 
-async function runJob(
+export const CRONS: Record<string, CronDef> = {
+  "Run daily jobs each day at 02h30": {
+    name: "Run daily jobs each day at 02h30",
+    cron_string: "30 2 * * *",
+    handler: async () => {
+      const { stdout, stderr } = await exec(
+        "/opt/app/run-dummy-outside-job.sh"
+      );
+      logger.info("stdout:", stdout);
+      logger.error("stderr:", stderr);
+      return stderr ? 1 : 0;
+    },
+  },
+  // "Run every 2 minutes dummy": {
+  //   name: "Run every 2 minutes dummy",
+  //   cron_string: "*/2 * * * *",
+  //   handler: async () => {
+  //     logger.info(`Dummy 2 minutes`);
+  //     return 0;
+  //   },
+  // },
+};
+
+export async function runJob(
   job: IJob,
   options: { runningLogs: boolean } = {
     runningLogs: true,
@@ -54,6 +59,9 @@ async function runJob(
   return executeJob(
     job,
     async () => {
+      if (job.type === "cron_task") {
+        return CRONS[job.name].handler();
+      }
       switch (job.name) {
         case "seed":
           return seed();
@@ -85,6 +93,14 @@ async function runJob(
         case "migrations:create":
           // eslint-disable-next-line @typescript-eslint/no-explicit-any
           return createMigration(job.payload as any);
+        case "crons:init": {
+          await cronsInit();
+          return;
+        }
+        case "crons:scheduler":
+          return cronsScheduler();
+
+        // BELOW SPECIFIC TO PRODUCT
         case "import:document":
           // eslint-disable-next-line @typescript-eslint/no-explicit-any
           return handleDocumentFileContent(job.payload as any);
@@ -97,23 +113,4 @@ async function runJob(
     },
     options
   );
-}
-
-//abortSignal
-export async function processor(): Promise<void> {
-  logger.info(`Process jobs queue - looking for a job to execute`);
-  const { value: nextJob } = await getDbCollection("jobs").findOneAndUpdate(
-    { status: "pending", scheduled_at: { $lte: new Date() } },
-    { $set: { status: "will_start" } },
-    { sort: { scheduled_at: 1 } }
-  );
-
-  if (nextJob) {
-    logger.info(`Process jobs queue - job ${nextJob.name} will start`);
-    await runJob(nextJob);
-  } else {
-    await sleep(60000); // 1 min
-  }
-
-  return processor();
 }